+++ app/javascript/__tests__/util/formEngineUtil.test.ts
@@ -1,5 +1,4 @@
 import dayjs from "dayjs"
-import MockDate from "mockdate"
 import {
   translationFromDataSchema,
   showStep,
@@ -9,7 +8,6 @@ import {
   validMonthRange,
   validYearRange,
   validDate,
-  validAge,
   parseDate,
   updateFormPath,
 } from "../../util/formEngineUtil"
@@ -249,33 +247,6 @@ describe("formEngineUtil", () => {
     })
     it("returns false for a valid date", () => {
       expect(validDate("2001", "2", "29")).toBe(false)
-    })
-  })
-
-  describe("validAge", () => {
-    MockDate.set("2020-01-01")
-    it("returns true for birth dates greater than minimum age", () => {
-      const birthDate = dayjs("2000-01-01")
-      const minimumAge = 18
-      expect(validAge(birthDate, minimumAge, undefined)).toBe(true)
-    })
-    it("returns false for birth dates less than minimum age", () => {
-      const birthDate = dayjs("2010-01-01")
-      const minimumAge = 18
-      expect(validAge(birthDate, minimumAge, undefined)).toBe(false)
-    })
-    it("returns true for birth dates that are less than 10 months in the futue", () => {
-      const birthDate = dayjs("2020-09-01")
-      expect(validAge(birthDate, null, undefined)).toBe(true)
-    })
-    it("overrides minimum age with senior building age requirements", () => {
-      const birthDate = dayjs("2000-01-01")
-      const minimumAge = 18
-      const seniorBuildingAgeRequirement = {
-        entireHousehold: true,
-        minimumAge: 65,
-      }
-      expect(validAge(birthDate, minimumAge, seniorBuildingAgeRequirement)).toBe(false)
     })
   })
